pyproject.toml
[project]
name = "ComfyUI_MiniCPM-V-2_6-int4"
description = "Implementation of [OpenBMB/MiniCPM-V](https://github.com/OpenBMB/MiniCPM-V) in ComfyUI"
version = "1.0.0"
license = "LICENSE"
dependencies = ["torch", "torchvision", "numpy", "pillow", "huggingface_hub", "transformers", "decord"]

[project.urls]
Repository = "https://github.com/IuvenisSapiens/ComfyUI_MiniCPM-V-2_6-int4"

[tool.comfy]
PublisherId = "IuvenisSapiens"
DisplayName = "ComfyUI_MiniCPM-V-2_6-int4"
Icon = "favicon.ico"
